api-server/requirements.txt
fastapi
uvicorn[standard]
pydantic
redis
httpx
python-multipart
python-dotenv
aiofiles